pyproject.toml
[project]
name = "nexus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "einops>=0.8.2",
    "faiss-cpu==1.10.0",
    "fastapi==0.115.12",
    "gitpython==3.1.44",
    "ipykernel>=7.2.0",
    "langchain==0.3.25",
    "langchain-community==0.3.23",
    "langchain-ollama>=0.3.10",
    "langgraph==0.4.1",
    "logging>=0.4.9.6",
    "matplotlib>=3.10.9",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "rich==14.0.0",
    "seaborn>=0.13.2",
    "sentence-transformers==4.1.0",
    "sqlalchemy==2.0.40",
    "typer==0.15.2",
    "uvicorn==0.34.2",
    "watchdog==6.0.0",
]
